Move application icon to their own module, to be able to have another icon for Element Enterprise
