+++ appicon/element/src/main/res/mipmap-xxxhdpi/ic_launcher_monochrome.webp
