+++ appicon/element/src/nightly/res/drawable/ic_launcher_background.xml
