+++ appicon/element/src/release/res/drawable/ic_launcher_background.xml
